requirements.txt
fastapi
uvicorn
mysql-connector-python
google-genai
python-dotenv
pydantic
